+++ CHANGELOG.md
@@ -87,6 +87,7 @@ All notable changes to this project will be documented in this file.
 - Provision build toolchain in prod image and restore before build ([#185](https://github.com/stevenfackley/StackAlchemist/issues/185))
 - Apply Supabase migrations in the prod deploy pipeline ([#186](https://github.com/stevenfackley/StackAlchemist/issues/186))
 - Require authentication and ownership in retryGeneration ([#189](https://github.com/stevenfackley/StackAlchemist/issues/189))
+- Make prod migration gate skip (not brick deploys) when secret unset ([#191](https://github.com/stevenfackley/StackAlchemist/issues/191))
 
 ### Dependencies
 
@@ -189,6 +190,7 @@ All notable changes to this project will be documented in this file.
 - Error category rendering (F6) ([#150](https://github.com/stevenfackley/StackAlchemist/issues/150))
 - Dashboard live updates island (F7) ([#151](https://github.com/stevenfackley/StackAlchemist/issues/151))
 - Adopt mise for task running and Node versioning ([#154](https://github.com/stevenfackley/StackAlchemist/issues/154))
+- Automatic Stripe refund when the compile guarantee fails ([#190](https://github.com/stevenfackley/StackAlchemist/issues/190))
 
 ### Miscellaneous Tasks